requirements.txt
attrs==23.1.0
aws-cdk-lib==2.100.0
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.155
aws-cdk.asset-node-proxy-agent-v6==2.0.1
cattrs==22.2.0
colorama==0.4.6
constructs==10.2.32
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.2
jsii==1.90.0
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.0.0
publication==0.0.3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.6.2
